requirements.txt
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
mauve-text>=0.3.0
torch>=2.1
transformers>=4.35
datasets>=2.14
accelerate>=0.24
pandas>=2.0
tqdm
